reverting script changes and added validator
